% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringilidae.R
\docType{data}
\name{dquery}
\alias{dquery}
\title{Fringillidae \code{datelifeQuery}}
\format{
A \code{datelifeQuery} object
}
\usage{
dquery
}
\description{
Fringillidae \code{datelifeQuery}
}
\details{
Generated with
dquery = datelife::make_datelife_query(input = "fringilidae", get_spp_from_taxon = TRUE)
usethis::use_data(dquery, overwrite = TRUE)
}
\keyword{datelife}
\keyword{fringillidae}
\keyword{tnrs}
