% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringilidae.R
\docType{data}
\name{dres}
\alias{dres}
\title{Fringillidae \code{datelifeResult}}
\format{
A \code{datelifeResult} object
}
\usage{
dres
}
\description{
Fringillidae \code{datelifeResult}
}
\details{
Generated with
dres = datelife::get_datelife_result(input = dquery, cache = opentree_chronograms)
usethis::use_data(dres, overwrite = TRUE)
}
\keyword{datelife}
\keyword{fringillidae}
\keyword{tnrs}
