% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringilidae.R
\docType{data}
\name{dsumm}
\alias{dsumm}
\title{Fringillidae \code{datelifeResultSummary}}
\format{
A \code{datelifeResultSummary} object
}
\usage{
dsumm
}
\description{
Fringillidae \code{datelifeResultSummary}
}
\details{
Generated with
dsumm = datelife::summary.datelifeResult(datelife_query = dquery, object = dres)
usethis::use_data(dsumm, overwrite = TRUE)
}
\keyword{datelife}
\keyword{fringillidae}
\keyword{tnrs}
